requirements.txt
fastapi
pdfminer.six
